Storage/Parallels.cpp: Fix harmless parfait integer overflow warning which is impossible, bugref:3409
svn:sync-xref-src-repo-rev: r166476

+++ src/VBox/Storage/Parallels.cpp
@@ -1,4 +1,4 @@
-/* $Id: Parallels.cpp 106061 2024-09-16 14:03:52Z knut.osmundsen@oracle.com $ */
+/* $Id: Parallels.cpp 107482 2025-01-06 16:32:07Z alexander.eichner@oracle.com $ */
 /** @file
  *
  * Parallels hdd disk image, core code.
@@ -695,7 +695,7 @@ static DECLCALLBACK(int) parallelsWrite(void *pBackendData, uint64_t uOffset, si
             Assert(uSector == 0);
             AssertMsg(pImage->cbFileCurrent % 512 == 0, ("File size is not a multiple of 512\n"));
             pImage->pAllocationBitmap[iIndexInAllocationTable] = (uint32_t)(pImage->cbFileCurrent / 512);
-            pImage->cbFileCurrent += pImage->PCHSGeometry.cSectors * 512;
+            pImage->cbFileCurrent += (uint64_t)pImage->PCHSGeometry.cSectors * 512;
             pImage->fAllocationBitmapChanged = true;
             uOffsetInFile = (uint64_t)pImage->pAllocationBitmap[iIndexInAllocationTable] * 512;
